requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
joblib==1.4.2
nltk==3.9.1
numpy==2.2.1
pandas==2.2.3
pyasn1==0.6.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
textblob==0.18.0.post0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
vaderSentiment==3.3.2
